+++ src/sc_keeper/parameters.py
@@ -5,7 +5,6 @@ from sc_crawler.table_fields import (
     Allocation,
     CpuAllocation,
     CpuArchitecture,
-    Status,
     StorageType,
     TrafficDirection,
 )
@@ -446,18 +445,6 @@ benchmark_config = Annotated[
     Optional[str],
     Query(
         description="Optional benchmark config dict JSON to filter results of a benchmark_id."
-    ),
-]
-
-status = Annotated[
-    Optional[Status],
-    Query(
-        title="Status",
-        description="Filter by resource status.",
-        json_schema_extra={
-            "category_id": FilterCategories.STATUS,
-            "enum": [s.value for s in Status],
-        },
     ),
 ]